# Specialize _memory_offset on number of arguments.