@KristofferC 2020-09-07
Seems like a pretty bad performance bug so I think it makes sense to backport htis